BIT/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Deprecated==1.2.9
Django==3.0.5
djangorestframework==3.10.2
idna==2.9
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
PyGithub==1.47
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
wrapt==1.12.1
